Moved the CollectionTests and RandomExtensionsTests to their own test classes
